pyproject.toml
[tool.poetry]
name = "fastapi-auth-cookiecutter"
version = "0.1.0"
description = ""
authors = ["SebaRossi94 <seba.rossi.94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
alembic = "^1.13.1"
SQLAlchemy = "^2.0.27"
sqlalchemy-utils = "^0.41.1"
sqlmodel = "^0.0.16"
pydantic-settings = "^2.2.1"
pydantic = {extras = ["email"], version = "^2.6.3"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
